refactor(sessions): move parent-session forking behind the accessor boundary
