+++ scripts/lib/session-accessor-debt-baseline.json
@@ -48,13 +48,12 @@
     "src/agents/subagent-spawn.ts": 2,
     "src/auto-reply/reply/commands-name.ts": 2,
     "src/auto-reply/reply/dispatch-from-config.ts": 2,
-    "src/auto-reply/reply/session-fork.ts": 2,
     "src/commands/doctor-heartbeat-main-session-repair.ts": 2,
     "src/commands/doctor-session-state-providers.ts": 2,
     "src/commands/doctor-state-integrity.ts": 2,
     "src/commands/doctor/shared/codex-route-warnings.ts": 2,
     "src/config/sessions/plugin-host-cleanup.ts": 2,
-    "src/config/sessions/session-accessor.ts": 13,
+    "src/config/sessions/session-accessor.ts": 14,
     "src/config/sessions/session-file.ts": 2,
     "src/config/sessions/session-registry-maintenance.ts": 2,
     "src/gateway/server-methods/agent.ts": 2,
