+++ src/auto-reply/reply/session-fork.test.ts
@@ -7,7 +7,6 @@ import type { OpenClawConfig } from "../../config/types.openclaw.js";
 import { forkSessionEntryFromParent } from "./session-fork.js";
 
 const runtimeMocks = vi.hoisted(() => ({
-  forkSessionFromParentRuntime: vi.fn(),
   resolveParentForkTokenCountRuntime: vi.fn(),
 }));
 
@@ -16,13 +15,14 @@ vi.mock("./session-fork.runtime.js", () => runtimeMocks);
 const roots: string[] = [];
 
 async function makeRoot(prefix: string): Promise<string> {
-  const root = await fs.mkdtemp(path.join(os.tmpdir(), prefix));
+  // realpath first: macOS tmpdir is a /var -> /private/var symlink and the
+  // fork resolver returns canonical paths.
+  const root = await fs.realpath(await fs.mkdtemp(path.join(os.tmpdir(), prefix)));
   roots.push(root);
   return root;
 }
 
 afterEach(async () => {
-  runtimeMocks.forkSessionFromParentRuntime.mockReset();
   runtimeMocks.resolveParentForkTokenCountRuntime.mockReset();
   await Promise.all(roots.splice(0).map((root) => fs.rm(root, { recursive: true, force: true })));
 });
@@ -38,13 +38,34 @@ describe("forkSessionEntryFromParent", () => {
     const configStorePath = path.join(configStoreDir, "sessions.json");
     const parentSessionKey = "agent:main:main";
     const sessionKey = "agent:main:subagent:child";
+    const parentSessionFile = path.join(activeStoreDir, "parent.jsonl");
+    await fs.writeFile(
+      parentSessionFile,
+      [
+        JSON.stringify({
+          type: "session",
+          version: 3,
+          id: "parent-session",
+          timestamp: "2026-05-01T00:00:00.000Z",
+          cwd: root,
+        }),
+        JSON.stringify({
+          type: "message",
+          id: "assistant-1",
+          parentId: null,
+          timestamp: "2026-05-01T00:00:01.000Z",
+          message: { role: "assistant", content: "hi" },
+        }),
+      ].join("\n") + "\n",
+      "utf-8",
+    );
     await fs.writeFile(
       storePath,
       JSON.stringify(
         {
           [parentSessionKey]: {
             sessionId: "parent-session",
-            sessionFile: path.join(activeStoreDir, "parent.jsonl"),
+            sessionFile: parentSessionFile,
             updatedAt: 1,
           },
           [sessionKey]: { sessionId: "", updatedAt: 2 },
@@ -56,12 +77,6 @@ describe("forkSessionEntryFromParent", () => {
     );
 
     runtimeMocks.resolveParentForkTokenCountRuntime.mockResolvedValue(10);
-    runtimeMocks.forkSessionFromParentRuntime.mockImplementation(
-      async ({ sessionsDir }: { sessionsDir: string }) => ({
-        sessionId: "forked-session",
-        sessionFile: path.join(sessionsDir, "forked-session.jsonl"),
-      }),
-    );
 
     const result = await forkSessionEntryFromParent({
       agentId: "main",
@@ -73,15 +88,16 @@ describe("forkSessionEntryFromParent", () => {
     });
 
     expect(result.status).toBe("forked");
-    expect(runtimeMocks.forkSessionFromParentRuntime).toHaveBeenCalledWith(
-      expect.objectContaining({
-        sessionsDir: activeStoreDir,
-      }),
-    );
+    if (result.status !== "forked") {
+      throw new Error("expected forked result");
+    }
+    // The fork artifact lands beside the store being mutated, not the config store.
+    expect(path.dirname(result.fork.sessionFile)).toBe(activeStoreDir);
     const stored = JSON.parse(await fs.readFile(storePath, "utf-8")) as Record<
       string,
-      { sessionFile?: string }
+      { sessionId?: string; sessionFile?: string }
     >;
-    expect(stored[sessionKey]?.sessionFile).toBe(path.join(activeStoreDir, "forked-session.jsonl"));
+    expect(stored[sessionKey]?.sessionId).toBe(result.fork.sessionId);
+    expect(stored[sessionKey]?.sessionFile).toBe(result.fork.sessionFile);
   });
 });
