+++ src/auto-reply/reply/session-parent-fork-prepare.ts
@@ -1,5 +1,4 @@
 // Prepares parent-context fork metadata for guarded reply session initialization.
-import path from "node:path";
 import type { SessionEntry } from "../../config/sessions.js";
 import { forkSessionFromParent, resolveParentForkDecision } from "./session-fork.js";
 
@@ -41,7 +40,9 @@ export async function prepareReplySessionParentFork(params: {
   const fork = await forkSessionFromParent({
     parentEntry,
     agentId: params.agentId,
-    sessionsDir: path.dirname(params.storePath),
+    parentSessionKey: params.parentSessionKey,
+    sessionKey: params.sessionKey,
+    storePath: params.storePath,
   });
   if (!fork) {
     return params.sessionEntry;
