+++ src/gateway/session-create-service.ts
@@ -1,5 +1,4 @@
 import { randomUUID } from "node:crypto";
-import path from "node:path";
 import {
   normalizeOptionalLowercaseString,
   normalizeOptionalString,
@@ -449,9 +448,11 @@ export async function createGatewaySession(params: {
         const fork = await forkSessionFromParent({
           parentEntry: currentParentSessionEntry,
           agentId: parentSessionTarget.agentId,
-          sessionsDir: path.dirname(parentSessionTarget.storePath),
+          parentSessionKey: canonicalParentSessionKey,
+          sessionKey: target.canonicalKey,
+          storePath: parentSessionTarget.storePath,
           // Keep the fork transcript owned by the child store across agent boundaries.
-          targetSessionsDir: path.dirname(target.storePath),
+          targetStorePath: target.storePath,
         });
         if (!fork) {
           return {
